Propagate the move of the manifest package to be public
