+++ pkg/sy/sy.go
@@ -20,10 +20,10 @@ import (
 
 	"github.com/gvallee/go_util/pkg/util"
 	"github.com/gvallee/kv/pkg/kv"
-	"github.com/sylabs/singularity-mpi/internal/pkg/manifest"
 	"github.com/sylabs/singularity-mpi/pkg/buildenv"
 	"github.com/sylabs/singularity-mpi/pkg/checker"
 	"github.com/sylabs/singularity-mpi/pkg/implem"
+	"github.com/sylabs/singularity-mpi/pkg/manifest"
 	"github.com/sylabs/singularity-mpi/pkg/syexec"
 	"github.com/sylabs/singularity-mpi/pkg/sys"
 )
